move strip tags to helpers
