+++ packages/chord-mark-converters/src/helpers/stripTags.js
